feat: implement edit description screen
