+++ src/constants/screens.ts
@@ -27,4 +27,5 @@ export enum Screen {
   MyProfileScreen = 'MyProfileScreen',
   EditProfileScreen = 'EditProfileScreen',
   SettingScreen = 'SettingScreen',
+  EditDescriptionScreen = 'EditDescriptionScreen',
 }
