+++ src/screens/FloorScreen/index.tsx
@@ -1,7 +1,14 @@
+import EditDescriptionScreen, {
+  EditDescriptionScreenParams,
+} from './EditDescriptionScreen';
 import EditFloorScreen, { EditFloorScreenParams } from './EditFloorScreen';
 import FloorViewerScreen, {
   FloorViewerScreenParams,
 } from './FloorViewerScreen';
 
-export { FloorViewerScreen, EditFloorScreen };
-export type { FloorViewerScreenParams, EditFloorScreenParams };
+export { FloorViewerScreen, EditFloorScreen, EditDescriptionScreen };
+export type {
+  FloorViewerScreenParams,
+  EditFloorScreenParams,
+  EditDescriptionScreenParams,
+};
